requirements.txt
Django==2.1.7
python-ldap==3.2.0
django-auth-ldap==1.7.0
django-ldapdb==1.3.0
Jinja2==2.10
Pillow==2.2.1